feat: implement update and delete to user

+++ app/Http/Controllers/Admin/UserController.php
@@ -15,10 +15,15 @@ class UserController extends Controller
         if ($request->ajax()) {
             return DataTables::of(User::latest()->get())
                 ->addColumn('action', function($data) {
+                    $user_method = [
+                        "PUT", 
+                        "/users/" . $data->id
+                    ];
+
                     return '
                         <div class="text-center">
                             <a href="' . route('users.show', $data->id) . '" class="btn btn-primary btn-sm"><i class="fas fa-folder"></i> View</a>
-                            <a href="javascript:void(0);" class="btn btn-info btn-sm"><i class="fas fa-pencil-alt"></i> Edit</a>
+                            <a href="javascript:void(0);" onclick=formModalType('. json_encode($user_method) .') class="btn btn-info btn-sm"><i class="fas fa-pencil-alt"></i> Edit</a>
                             <a href="javascript:void(0);" onclick=deleteUser(' . $data->id . ') class="btn btn-danger btn-sm"><i class="fas fa-trash"></i> Delete</a>
                         </div>
                     ';
@@ -49,16 +54,38 @@ class UserController extends Controller
         return $user;
     }
 
-    public function update(Request $request, User $user)
+    public function update(UserRequest $request, $user_id)
     {
-        //
+        $data = [
+            'request' => $request,
+            'user_id' => $user_id,
+            'method' => 'update'
+        ];
+        return $this->isUpdateOrDestroy($data);
     }
 
     public function destroy($id)
     {
-        $user = User::find($id);
+        $data = [
+            'user_id' => $id,
+            'method' => 'destroy'
+        ];
+        return $this->isUpdateOrDestroy($data);
+    }
 
-        if($user) {
+    private function isUpdateOrDestroy($data) {
+        $user = User::find($data['user_id']);
+        
+        if($user && ($data['method'] == 'update' || $data['method'] == 'patch')) {
+            $user->update($data['request']->validated());
+
+            return response()->json([
+                'title' => 'Updated', 
+                'text' => 'User successfully updated', 
+                'icon' => 'success',
+            ], 200);
+        }
+        else if($user && $data['method'] == 'destroy') {
             $user->delete();
             
             return response()->json([
@@ -66,7 +93,7 @@ class UserController extends Controller
                 'text' => 'User has been deleted', 
                 'icon' => 'success'
             ], 200);
-        } 
+        }
 
         return response()->json([
             'title' => 'Failed',
